add and improve code for notify
