+++ Cargo.toml
@@ -31,6 +31,7 @@ members = [
     "crates/obs", # Observability utilities
     "crates/protos", # Protocol buffer definitions
     "crates/rio", # Rust I/O utilities and abstractions
+    "crates/targets", # Target-specific configurations and utilities
     "crates/s3select-api", # S3 Select API interface
     "crates/s3select-query", # S3 Select query engine
     "crates/signer", # client signer
@@ -86,6 +87,7 @@ rustfs-signer = { path = "crates/signer", version = "0.0.5" }
 rustfs-checksums = { path = "crates/checksums", version = "0.0.5" }
 rustfs-workers = { path = "crates/workers", version = "0.0.5" }
 rustfs-mcp = { path = "crates/mcp", version = "0.0.5" }
+rustfs-targets = { path = "crates/targets", version = "0.0.5" }
 aes-gcm = { version = "0.10.3", features = ["std"] }
 anyhow = "1.0.99"
 arc-swap = "1.7.1"
@@ -93,15 +95,15 @@ argon2 = { version = "0.5.3", features = ["std"] }
 atoi = "2.0.0"
 async-channel = "2.5.0"
 async-recursion = "1.1.1"
-async-trait = "0.1.88"
+async-trait = "0.1.89"
 async-compression = { version = "0.4.19" }
 atomic_enum = "0.3.0"
-aws-config = { version = "1.8.4" }
+aws-config = { version = "1.8.5" }
 aws-sdk-s3 = "1.101.0"
 axum = "0.8.4"
 base64-simd = "0.8.0"
 base64 = "0.22.1"
-brotli = "8.0.1"
+brotli = "8.0.2"
 bytes = { version = "1.10.1", features = ["serde"] }
 bytesize = "2.0.1"
 byteorder = "1.5.0"
@@ -109,7 +111,7 @@ cfg-if = "1.0.1"
 crc-fast = "1.4.0"
 chacha20poly1305 = { version = "0.10.1" }
 chrono = { version = "0.4.41", features = ["serde"] }
-clap = { version = "4.5.44", features = ["derive", "env"] }
+clap = { version = "4.5.45", features = ["derive", "env"] }
 const-str = { version = "0.6.4", features = ["std", "proc"] }
 crc32fast = "1.5.0"
 criterion = { version = "0.7", features = ["html_reports"] }
@@ -118,7 +120,7 @@ datafusion = "46.0.1"
 derive_builder = "0.20.2"
 dioxus = { version = "0.6.3", features = ["router"] }
 dirs = "6.0.0"
-enumset = "1.1.7"
+enumset = "1.1.9"
 flatbuffers = "25.2.10"
 flate2 = "1.1.2"
 flexi_logger = { version = "0.31.2", features = ["trc", "dont_minimize_extra_stacks"] }
@@ -131,7 +133,7 @@ hex = "0.4.3"
 hex-simd = "0.8.0"
 highway = { version = "1.3.0" }
 hmac = "0.12.1"
-hyper = "1.6.0"
+hyper = "1.7.0"
 hyper-util = { version = "0.1.16", features = [
     "tokio",
     "server-auto",
@@ -190,7 +192,7 @@ rand = "0.9.2"
 rdkafka = { version = "0.38.0", features = ["tokio"] }
 reed-solomon-simd = { version = "3.0.1" }
 regex = { version = "1.11.1" }
-reqwest = { version = "0.12.22", default-features = false, features = [
+reqwest = { version = "0.12.23", default-features = false, features = [
     "rustls-tls",
     "charset",
     "http2",
@@ -217,7 +219,7 @@ rustls-pemfile = "2.2.0"
 s3s = { version = "0.12.0-minio-preview.3" }
 schemars = "1.0.4"
 serde = { version = "1.0.219", features = ["derive"] }
-serde_json = { version = "1.0.142", features = ["raw_value"] }
+serde_json = { version = "1.0.143", features = ["raw_value"] }
 serde_urlencoded = "0.7.1"
 serial_test = "3.2.0"
 sha1 = "0.10.6"
@@ -234,7 +236,7 @@ sysctl = "0.6.0"
 tempfile = "3.20.0"
 temp-env = "0.3.6"
 test-case = "3.3.1"
-thiserror = "2.0.14"
+thiserror = "2.0.15"
 time = { version = "0.3.41", features = [
     "std",
     "parsing",
