+++ crates/notify/src/rules/rules_map.rs
@@ -14,8 +14,8 @@
 
 use super::pattern_rules::PatternRules;
 use super::target_id_set::TargetIdSet;
-use crate::arn::TargetID;
-use crate::event::EventName;
+use rustfs_targets::arn::TargetID;
+use rustfs_targets::EventName;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
