+++ crates/notify/src/rules/target_id_set.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::arn::TargetID;
+use rustfs_targets::arn::TargetID;
 use std::collections::HashSet;
 
 /// TargetIDSet - A collection representation of TargetID.
