+++ crates/notify/src/rules/xml_config.rs
@@ -13,8 +13,8 @@
 // limitations under the License.
 
 use super::pattern;
-use crate::arn::{ARN, ArnError, TargetIDError};
-use crate::event::EventName;
+use rustfs_targets::arn::{ArnError, TargetIDError, ARN};
+use rustfs_targets::EventName;
 use serde::{Deserialize, Serialize};
 use std::collections::HashSet;
 use std::io::Read;
